Implement processClient method in PasswordServer
Worth noticing is that the getCredentials method is part of the "UserAccount events", since it takes a user account as a parameter. Also the getAccount method is currently using null values as parameters. I do not want to implement this as a special event type, so we should be able to use the UserAccount class to implement this method as well. The parameter values are null for now since the UserAccount class is not yet implemented.