Convert Eclipse project to Maven project
Using a Maven project will simplify the use of external libraries, such as SQLite which is used in this project. Also simplifies further use of external libraries. Building, testing and deployments can be automated in Github every time code is merged into the master branch, and/or when a pull request is created.