Add support for multiple encryption methods
The CommunicationProtocol will use different encryption methods depending on what operation it is doing. No encryption is used initially to exchange RSA public keys between the client and server. Later on in the program, when the RSA keys have been exchanged, the RSA encryption will be used to exchange an AES key from the server to the client. When the AES key has been generated on the server and exchanged with the client, the AES encryption will be used. The AES key will be continuously changed throughout the lifetime of the connection. This is handled automatically by the CommunicationProtocol.