Replace Multimap with custom Message class
The CommunicationProtocol was very complicated to use before. With the addition of a generic Message class, and Query, and Response classes that inherit from this abstract class, the code is much cleaner and we can receive more data from the server such as a response code that tells us if the query failed or not.